dashboard/requirements.txt
# Enterprise Surveillance Dashboard — Plotly Dash (2026)
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
# SQLite via stdlib; no extra pkg
